refactor(ext/node): delete _fs_common.ts, move callback helpers into internal/fs/utils.mjs
The shared callback-form helpers (callbackify/callbackifyOpt/callbackifyWrite
with their FSReqCallback tracking, makeCallback, encoding/signal option
probes) move into internal/fs/utils.mjs -- node's actual home for shared fs
helpers, and already an eager import of every consumer. The type-only imports
in handle.ts/promises.ts pointed at type exports that never existed (erased
to any); they're now explicit local aliases. _fs/ is down to _fs_glob.ts.
