+++ ext/node/lib.rs
@@ -801,7 +801,6 @@ deno_core::extension!(deno_node,
     "string_decoder.ts",
     "testing.ts",
     "test/reporters.ts",
-    "_fs/_fs_common.ts",
     "path/_posix.ts",
     "path/_win32.ts",
     "path/mod.ts",
