+++ tools/core_import_map.json
@@ -528,7 +528,6 @@
     "ext:deno_net/02_tls.js": "../ext/net/02_tls.js",
     "ext:deno_node/_events.d.ts": "../ext/node/polyfills/_events.d.ts",
     "ext:deno_node/internal/fs/promises.ts": "../ext/node/polyfills/internal/fs/promises.ts",
-    "ext:deno_node/_fs/_fs_common.ts": "../ext/node/polyfills/_fs/_fs_common.ts",
     "ext:deno_node/_fs/_fs_fchown.ts": "../ext/node/polyfills/_fs/_fs_fchown.ts",
     "ext:deno_node/_fs/_fs_futimes.ts": "../ext/node/polyfills/_fs/_fs_futimes.ts",
     "ext:deno_node/_fs/_fs_ftruncate.ts": "../ext/node/polyfills/_fs/_fs_ftruncate.ts",
